requirements.txt
# --- Server ---
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pydantic>=2.6.0
pydantic-settings>=2.2.0

# --- LLM ---
google-genai>=1.0.0
ollama>=0.2.0

# --- Shared ---
psutil>=5.9.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# --- CLI ---
typer>=0.12.0
rich>=13.7.0

# --- Dev ---
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
